requirements.txt
streamlit==1.32.0
opencv-python-headless==4.8.1.78
numpy==1.26.3
pillow==10.2.0
llama-index-llms-gemini==0.1.4
llama-index-core==0.10.11
google-generativeai==0.3.2
pandas==2.1.4
altair==5.2.0
python-dotenv==1.0.0
protobuf==4.25.1 